requirements.txt
typer>=0.12,<1.0
pydantic>=2.7,<3.0
faster-whisper>=1.0.3,<2.0
python-dotenv>=1.0,<2.0
